fix double in php generator

+++ samples/client/petstore/php/OpenAPIClient-php/lib/Model/FormatTest.php
@@ -62,7 +62,7 @@ class FormatTest implements ModelInterface, ArrayAccess, \JsonSerializable
         'int64' => 'int',
         'number' => 'float',
         'float' => 'float',
-        'double' => 'double',
+        'double' => 'float',
         'decimal' => 'float',
         'string' => 'string',
         'byte' => 'string',
@@ -636,7 +636,7 @@ class FormatTest implements ModelInterface, ArrayAccess, \JsonSerializable
     /**
      * Gets double
      *
-     * @return double|null
+     * @return float|null
      */
     public function getDouble()
     {
@@ -646,7 +646,7 @@ class FormatTest implements ModelInterface, ArrayAccess, \JsonSerializable
     /**
      * Sets double
      *
-     * @param double|null $double double
+     * @param float|null $double double
      *
      * @return self
      */
